rename file from store
