+++ 17_Global State Management and Data Fetching/pratikum/react-todos/src/store/store.js
@@ -1,7 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit";
 import { persistStore, persistReducer } from "redux-persist"
 import storage from "redux-persist/lib/storage";
-import reducer from "./passengerSlice";
+import reducer from "./todoSlice";
 
 const persistConfig = {
     key: "Kennedy",
