+++ 17_Global State Management and Data Fetching/pratikum/react-todos/src/store/todoSlice.js
